Take AutoSwitch into account.
********************************************************************************
Basically, during the section immediately after the first "Option:AutoSwitch",
the accounts are defined at a high level, including things like credit limits,
but without any transactions.  Then, there is a "Clear:AutoSwitch", and that ends
the section.  Then, later, the accounts are defined again, this time with
transactions, but possibly without some other stuff, like the credit limit.

So, we define the account the first time, and note that we were in the first
"auto switch" section.  Later, if we encounter an account with the same name,
we assume that this is "filling in" the transactions from the already-defined
account.  We verify this based on the "AutoSwitch" instances.  We only do a
merge if the initial definition was immediately following the first AutoSwitch,
and the next definition was NOT.

Previously, two accounts could be defined with the same name, and they would be
kept distinct.  This was likely a bug.  It makes no sense to have two accounts
with the same name.  Now, we will detect if they are the mergeable combination,
and if so, merge them; and, if not, it is an unrecoverable error.

The "counting" of the AutoSwitch'es here is hacky, but it seems to get the job
done.  Note that I have found very little documentation about AutoSwitch.
A lot of this is guessing / reverse-engineering.