Add a new test.
This test reads in a QIF file, and then, one by one, verifies all sorts of
expected values.  It is pretty pedantic, and is all inside a single test, so the
summary ("Ran 1 test") is not reflective of what it's doing.  There may be more
elegant ways to write this test.

The QIF file used in this test is not hand-written, but based on a real QIF file
created by Quicken.  Specifically, I used Quicken 2008 Deluxe for Windows, using
the QIF Export, with "<All Accounts>" selected, for all dates, and with
"Transactions", "Memorized Payees", "Account List", and "Category List" chosen
for export.  (Other checkboxes offered were "For Macintosh Users" and "Security
Lists"; I left these un-selected.)

A few things in particular that are found in this file, which were not handled
well by the initial version of this software that I am extending:
- Windows-style newlines (I'm running the QIF parse on MacOS X)
- amounts written with commas in them (100,000.00)
- dates in US-style m/d/yy format (previous test was European d/m)
- dates using format like "12/31' 2"; presumably this has something to do with
  y2k issues.  I believe this format is meant to indicate definitively that the
  date does not represent 1902, but 2002.
- line items in transactions using the 'U' key.  I really have no idea what this
  is supposed to mean.  In my experience, the 'U' value is always identical to
  the 'T' value.  It is simply repeated information.  Since that's what it
  appears to be, the test expects it to be so.
- the type for a credit card is written as 'CCard', with both "C"s capitalized;
  the program previously only supported 'Ccard', with the second "C" lower case
- this file defines a "Tag", which were not supported at all before
- non-leaf categories receive a 'ZZZZZ' subcategory; in other words, we define
  categories 'Auto' and 'Auto:Registration', but the QIF file defines 'Auto',
  'Auto:Registration', and ALSO defines 'Auto:ZZZZZ'.  Right now, the program
  does not handle this specially.  It actually defines a category as such.
  But the test "knows" about it, and pulls out the ZZZZZ's and handles them
  separately.

However, note that the QIF file being checked in with this commit is not exactly
the one that was generated by Quicken.  I needed to eliminate one line; at line
129 of the orginal file, there was:

!Type:CCard

It is immediately followed by "!Account" on the next line.  According to what
I've read about the format, this really doesn't make sense, and certainly
doesn't work here.  Currently the program breaks the input up into chunks, where
each chunk is delimited by a caret, on a line by itself.  This puts
"!Type:CCard" and "!Account" into a single chunk, which utterly confuses things.
For this commit, simply remove "!Type:CCard".