Add a default account type (Cash) and a test to recognize "obfuscated" types.
I don't know what causes these weird account types.  Maybe it's just a bug in
Quicken.  Or maybe they have some reason for "obfuscating" the type.  All I know
for now is that file I export have these weird types in them, and I want to be
able to parse them.

When the parser encounters a weird account type, it substitues the "default"
account, which is defined in qifparse/qif.

This isn't the only "type" problem I encountered.  With older QIF files, I found
accounts of type "Port", "Mutual", and "401(k)/403(b)".  In each case, the
account was also identified as being type "Invst", which makes sense, as each of
the previous types is a kind of investment.  So I added a new dictionary of
"subtypes" which maps to the official type of account that should be used.